fixing screen size stuff again

+++ main.cpp
@@ -16,9 +16,8 @@ int main (int argc, char *argv[]) {
 	HANDLE hConsole = GetStdHandle(STD_OUTPUT_HANDLE);
 	_CONSOLE_SCREEN_BUFFER_INFO info;	
 	GetConsoleScreenBufferInfo(hConsole,&info);
-	//SMALL_RECT windowSize = {0, 0, 119, 59};
-	int windowSizeX = info.dwMaximumWindowSize.X-3;
-	int windowSizeY = info.dwMaximumWindowSize.Y-3;
+	short int windowSizeX = info.dwMaximumWindowSize.X-3;
+	short int windowSizeY = info.dwMaximumWindowSize.Y-3;
 	SMALL_RECT windowMaxSize = {0, 0, windowSizeX, windowSizeY};
 	
 	SetConsoleWindowInfo(hConsole, TRUE, &windowMaxSize);
